fix: fix display of component when invalid but no error template
When there is no error to display because the component has an error, but none of the templates can display that error, the component is not displayed. This makes sure margins or borders that could exist on the component are invisible in that case.
Moreover, the style is not set to block when the component is displayed. It's simply not set, so that users can display inline errors rather than block errors if they want to.