refactor: extract snippets out of source files
This is not ideal, because it requires on node's require and on the raw webpack loader that is not officially supported by the CLI, but I still think it makes it better. If we really can't use that strategy in the long run, we can still go back to the previous solution.

I have the feeling that [this](https://www.typescriptlang.org/docs/handbook/modules.html#wildcard-module-declarations) would be a better solution, bu I couldn't make that work.