fix: force string conversion in prism
It looks like there is a difference in behavior that I don't really get, but without forcing the string conversion, all our tests are failing with "`code.trim` is not a function" because `code` contains an object and not a string (it is the result of the `require` calls to laod the dynamic snippets). I may be missing something obvious, but this fixes the issue.