feat: make the component OnPush
Further changes should be done to only pass parts of the validation state (dirty, touched, submitted) to `this.config.shouldDisplayErrors()`. And it's a breaking change, since other conditions would not work anymore.
Besides, I have no idea if this change actually makes a difference in terms of performance, but I think it should, because it doesn't recompute all the errors to display at each CD.
And I fear that it's not actually completely safe to do it this way, so maybe we should just wait until forms use signals.